moved service component to lib/services/ftp module, shared between paramiko and pysftp implementation
