+++ projects/gnrcore/packages/sys/lib/services/ftp.py
@@ -4,9 +4,11 @@ import re
 
 from datetime import datetime
 from gnr.core import gnrstring
-from gnr.core.gnrbag import Bag,DirectoryResolver,BagResolver
+from gnr.core.gnrbag import Bag, DirectoryResolver, BagResolver
 from gnr.lib.services import GnrBaseService
-
+from gnr.core.gnrdecorator import public_method
+from gnr.web.gnrbaseclasses import BaseComponent
+from gnr.web.gnrwebstruct import struct_method
 
 class SftpService(GnrBaseService):
     def __init__(self, parent=None,host=None,username=None,password=None,private_key=None,port=None):
@@ -114,3 +116,88 @@ class SftpDirectoryResolver(DirectoryResolver):
         return SftpDirectoryResolver(path,  os.path.basename(path), 
                                     **self.instanceKwargs)
         
+
+
+class SftpClient(BaseComponent):
+    py_requires='public:Public'
+
+    @struct_method
+    def sftp_sftpClientLayout(self,pane,ftpname=None,
+                            datapath='.sftpclient',destdir=None,remotedir=None,**kwargs):
+        bc = pane.borderContainer(datapath=datapath,_anchor=True,**kwargs)
+        self.sftp_remoteTree(bc.roundedGroupFrame(region='left',title='!!Remote',
+                            datapath='.remote',width='50%',
+                            splitter=True),ftpname=ftpname,remotedir=remotedir)
+        self.sftp_localTree(bc.roundedGroupFrame(region='center',title='!!Local',
+                            datapath='.local'),ftpname=ftpname,destdir=destdir)
+
+    def sftp_remoteTree(self,frame,ftpname=None,remotedir=None):
+        resolver = self.getService('ftp',ftpname).sftpResolver()
+        frame.data('.tree',resolver())
+        self.sftp_fileTree(frame,nodeId='%s_src' %ftpname,topic='%s_upload' %ftpname)
+        frame.dataRpc(None,self.sftp_uploadFiles,ftp=ftpname,
+                    _onResult="""kwargs._dropnode.refresh(true);""",
+                    **{'subscribe_%s_upload' %ftpname:True})
+
+    def sftp_localTree(self,frame,ftpname=None,destdir=None):
+        resolver= DirectoryResolver(destdir or self.site.getStatic('site').path())
+        frame.data('.tree',resolver())
+        self.sftp_fileTree(frame,nodeId='%s_dest' %ftpname,
+                            topic='%s_download' %ftpname)
+        frame.dataRpc(None,self.sftp_downloadFiles,ftp=ftpname,
+                    _onResult="""kwargs._dropnode.refresh(true);""",
+                        **{'subscribe_%s_download' %ftpname:True})
+
+
+    def sftp_onDrag(self):
+        return """var children=treeItem.getValue('static')
+                  if(!children){
+                      dragValues['fsource']=[treeItem.attr.abs_path];
+                      return
+                  }
+                   result=[];
+                   children.forEach(function(n){
+                        if (n.attr.checked && !n._value){result.push(n.attr.abs_path);
+                    }},'static');
+                   dragValues['fsource']= result; 
+               """
+
+    @public_method
+    def sftp_downloadFiles(self,sourcefiles=None,destfolder=None,ftp=None,**kwargs):
+        self.getService('ftp',ftp).downloadFilesIntoFolder(sourcefiles=sourcefiles,
+                                                destfolder=destfolder,**kwargs)
+
+    @public_method
+    def sftp_uploadFiles(self,sourcefiles=None,destfolder=None,ftp=None,**kwargs):
+        self.getService('ftp',ftp).uploadFilesIntoFolder(sourcefiles=sourcefiles,
+                                                destfolder=destfolder,**kwargs)
+
+
+    def sftp_fileTree(self,pane,topic=None,**kwargs):
+        tree = pane.treeGrid(storepath='.tree',hideValues=True, 
+                      selectedLabelClass='selectedTreeNode',
+                      selected_abs_path='.abs_path',selected_file_ext='.file_ext',
+                      checked_abs_path='.checked_abs_path',
+                      #labelAttribute='nodecaption',
+                       autoCollapse=True,
+                      onDrag_fsource=self.sftp_onDrag(),
+                      headers=True,draggable=True,dragClass='draggedItem',
+                      onDrop_fsource="""
+                         if(dropInfo.treeItem.attr.file_ext!='directory'){
+                             return false;
+                         }else{
+                             genro.publish('%s',{
+                                destfolder:dropInfo.treeItem.attr.abs_path,
+                                _dropnode:dropInfo.treeItem,
+                                sourcefiles:data});
+                         }
+                     """ %topic,dropTargetCb_fsource="""
+                     if(dropInfo.selfdrop || dropInfo.treeItem.attr.file_ext!='directory'){
+                         return false;
+                     }
+                     return true;
+                     """,**kwargs)
+        tree.column('nodecaption',header='!!Name')
+        tree.column('file_ext',size=50,header='!!Ext')
+        tree.column('size',header='!!Size(KB)',size=60,dtype='L')
+        tree.column('mtime',header='!!MTime',size=100,dtype='DH')
